unidad 2/freefire_api/requirements.txt
Django>=5.0
 djangorestframework
